[refactor] Move OpType methods back to their rightful place
